requirements.txt
fastavro==1.4.1
kafka-python==2.0.2
requests==2.25.1
